pyproject.toml
[project]
name = "stock-alert-bot"
version = "0.1.0"
description = "매일 주식 MDD와 Fear & Greed Index를 텔레그램으로 알려주는 봇"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "yfinance>=0.2.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "python-telegram-bot>=21.0",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
]
